requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
sentence-transformers
uvicorn
ctransformers
fastapi
ipykernel
python-box
unstructured
tiktoken
pinecone-client
pypdf
google-generativeai
PyPDF2
chromadb
